backend/app/requirements.txt
opentelemetry-distro==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-instrumentation-logging==0.45b0
prometheus-client==0.20.0
